+++ src/android/HealthPlugin.java
@@ -1041,11 +1041,6 @@ public class HealthPlugin extends CordovaPlugin {
                 return;
             }
 
-            if (!args.getJSONObject(0).has("value")) {
-                callbackContext.error("Missing argument value");
-                return;
-            }
-
             InsertRecordsResponse response;
             List<Record> data = new LinkedList<>();
 
@@ -1067,7 +1062,7 @@ public class HealthPlugin extends CordovaPlugin {
                 BodyFatRecord record = new BodyFatRecord(
                         Instant.ofEpochMilli(st), null,
                         new Percentage(perc),
-                        Metadata.EMPTY);
+                        Metadata.unknownRecordingMethod());
                 data.add(record);
             } else if (datatype.equalsIgnoreCase("activity")) {
                 String activityStr = args.getJSONObject(0).getString("value");
@@ -1080,9 +1075,9 @@ public class HealthPlugin extends CordovaPlugin {
                 ExerciseSessionRecord record = new ExerciseSessionRecord(
                         Instant.ofEpochMilli(st), null,
                         Instant.ofEpochMilli(et), null,
+                        Metadata.unknownRecordingMethod(),
                         exerciseType,
                         title, notes,
-                        Metadata.EMPTY,
                         segments, laps);
                 data.add(record);
             } else if (datatype.equalsIgnoreCase("calories")) {
@@ -1092,7 +1087,7 @@ public class HealthPlugin extends CordovaPlugin {
                         Instant.ofEpochMilli(st), null,
                         Instant.ofEpochMilli(et), null,
                         Energy.kilocalories(kcals),
-                        Metadata.EMPTY);
+                        Metadata.unknownRecordingMethod());
                 data.add(record);
             } else if (datatype.equalsIgnoreCase("calories.active")) {
                 double kcals = args.getJSONObject(0).getDouble("value");
@@ -1101,7 +1096,7 @@ public class HealthPlugin extends CordovaPlugin {
                         Instant.ofEpochMilli(st), null,
                         Instant.ofEpochMilli(et), null,
                         Energy.kilocalories(kcals),
-                        Metadata.EMPTY);
+                        Metadata.unknownRecordingMethod());
                 data.add(record);
             } else if (datatype.equalsIgnoreCase("calories.basal")) {
                 double kcals = args.getJSONObject(0).getDouble("value");
@@ -1113,7 +1108,7 @@ public class HealthPlugin extends CordovaPlugin {
                 BasalMetabolicRateRecord record = new BasalMetabolicRateRecord(
                         Instant.ofEpochMilli(st), null,
                         pow,
-                        Metadata.EMPTY);
+                        Metadata.unknownRecordingMethod());
                 data.add(record);
             } else if (datatype.equalsIgnoreCase("blood_glucose")) {
                 JSONObject glucoseobj = args.getJSONObject(0).getJSONObject("value");
@@ -1129,7 +1124,7 @@ public class HealthPlugin extends CordovaPlugin {
                         Instant.ofEpochMilli(st), null,
                         Instant.ofEpochMilli(et), null,
                         len,
-                        Metadata.EMPTY);
+                        Metadata.unknownRecordingMethod());
 
                 data.add(record);
             } else if (datatype.equalsIgnoreCase("sleep")) {