+++ src/android/StepsFunctions.java
@@ -75,7 +75,7 @@ public class StepsFunctions {
                 Instant.ofEpochMilli(st), null,
                 Instant.ofEpochMilli(et), null,
                 steps,
-                Metadata.EMPTY
+                Metadata.unknownRecordingMethod()
         );
         data.add(record);
     }